refactored the set command to have the same structure as the put and add commands
